Move `TypeData` associated functions to `src::types::fmt::util`
